Removed the remaining size_t references that affected the FIFE API and replaced them with more appropriate definitions.  The remaining size_t definitions will remain as they wont cause any issues.  fixes[t:152]

+++ engine/core/model/model.cpp
@@ -129,7 +129,7 @@ namespace FIFE {
 		}
 	}
 
-	size_t Model::getNumMaps() const {
+	uint32_t Model::getNumMaps() const {
 		return m_maps.size();
 	}
